Simplify scenario creation, log failed URLs
